feat(account): modified SidebarAccount.jsx and TabbarAccount.jsx
