+++ src/pages/OrderHistory.jsx
@@ -1,9 +1,18 @@
-import React from 'react'
+import React from 'react';
 
 const OrderHistory = () => {
-    return (
-        <div>OrderHistory</div>
-    )
+
+
+  return (
+    <main className="px-4 sm:w-4xl">
+        <h1 className="hidden mb-8 text-4xl font-bold sm:block">
+        การซื้อของฉัน
+        </h1>
+        <div className="w-full text-center bg-gray-200 rounded-lg h-100">
+            ตาราง
+        </div>
+    </main>
+)
 }
 
 export default OrderHistory
