+++ src/pages/Profile.jsx
@@ -5,7 +5,7 @@ import TabbarAccount from '../components/account/TabbarAccount'
 
 const Profile = () => {
   return (
-    <div className="flex justify-between min-h-[80dvh] mx-auto mt-8 max-w-7xl outline-1">
+    <div className="flex flex-col sm:flex-row sm:justify-between min-h-[80dvh] mx-auto mt-8 max-w-7xl ">
       <SidebarAccount />
       <TabbarAccount />
       <Outlet />
